- refactor timer api, fix some errors

+++ examples/code/dynamic_cubemap/dynamic_cubemap.cpp
@@ -98,8 +98,8 @@ void example_setup(ecs_scene* scene, camera& cam)
     scene->samplers[ground].sb[0].sampler_state = pmfx::get_render_state(PEN_HASH("clamp_linear"), pmfx::RS_SAMPLER);
     bake_material_handles(scene, ground);
 
-    f32 ramp1_angle[] = {M_PI * 0.125, M_PI * -0.125};
-    f32 ramp2_angle[] = {M_PI * -0.125, M_PI * 0.125};
+    f32 ramp1_angle[] = {(f32)M_PI * 0.125f, (f32)M_PI * -0.125f};
+    f32 ramp2_angle[] = { (f32)M_PI * -0.125f, (f32)M_PI * 0.125f};
     f32 ramp1_x[] = {20.0f, -20.0f};
     f32 ramp2_x[] = {-10.0f, 10.0f};
     f32 block1_x[] = {30.0f, -30.0f};
